backend/requirements.txt
# Flask and Extensions
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-migrate==4.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ML and Data Processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
werkzeug==3.0.1
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
email-validator==2.1.0.post1
redis==4.5.5
bleach==6.0.0
click>=8.1.3
rich>=13.3.0

# Server
gunicorn==21.2.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0
pre-commit==3.5.0
